End of Section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -60,12 +61,12 @@ public class AuctionsController: ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        //TODO: Add current user as seller
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
         _context.Auctions.Add(auction);
         var newAuction = _mapper.Map<AuctionDto>(auction);
         await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
@@ -79,6 +80,7 @@ public class AuctionsController: ControllerBase
             new {auction.Id}, _mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -88,7 +90,7 @@ public class AuctionsController: ControllerBase
 
         if (auction == null) return NotFound();
 
-        //TODO: Check sellet == username
+        if(auction.Seller != User.Identity.Name) return Forbid("");
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -106,6 +108,7 @@ public class AuctionsController: ControllerBase
 
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -113,7 +116,7 @@ public class AuctionsController: ControllerBase
 
         if (auction == null) return NotFound();
 
-        //TODO: check seller == user_name
+        if(auction.Seller != User.Identity.Name) return Forbid("");
 
         _context.Auctions.Remove(auction);
